backend/requirements/base.txt
Django>=5.2
psycopg2-binary
django-cors-headers
djangorestframework
djangorestframework-simplejwt